Move business route to bottom change to slug
This commit is actually really important because before
this anytime a user would visit a route like login or
logout or sign in it would immediately go to business
show and think that login was the business name. Moving
it to the bottom fixes this problem. Also we renamed it
to slug to keep it consistent in the params hash.