Adds Job feature test to create a job with address
In order to create a job a user should have to enter an address.
This test is testing many aspects of our application together
The first thing it tests is that a business can create a job.
It then tests whether a job is created with an address if it is
saved under job as a relationship.